vacuum/alarm/requirements.txt
pyverisure==2.6.1
